refactor: move composables to ui folder and separate by utility
