+++ app/src/main/java/com/github/onlynotesswent/ui/overview/Overview.kt
@@ -34,15 +34,15 @@ import com.github.onlynotesswent.model.folder.FolderViewModel
 import com.github.onlynotesswent.model.note.Note
 import com.github.onlynotesswent.model.note.NoteViewModel
 import com.github.onlynotesswent.model.users.UserViewModel
+import com.github.onlynotesswent.ui.common.CustomDropDownMenu
+import com.github.onlynotesswent.ui.common.CustomDropDownMenuItem
+import com.github.onlynotesswent.ui.common.CustomLazyGrid
+import com.github.onlynotesswent.ui.common.FolderDialog
 import com.github.onlynotesswent.ui.navigation.BottomNavigationMenu
 import com.github.onlynotesswent.ui.navigation.LIST_TOP_LEVEL_DESTINATION
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.github.onlynotesswent.ui.navigation.Screen
 import com.github.onlynotesswent.ui.navigation.TopLevelDestinations
-import com.github.onlynotesswent.utils.CustomDropDownMenu
-import com.github.onlynotesswent.utils.CustomDropDownMenuItem
-import com.github.onlynotesswent.utils.CustomLazyGrid
-import com.github.onlynotesswent.utils.FolderDialog
 
 /**
  * Displays the overview screen which contains a list of publicNotes retrieved from the ViewModel.
